@@ -18,18 +18,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.0.0",
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
+    "@angular/animations": "^4.3.4",
+    "@angular/common": "^4.3.4",
+    "@angular/compiler": "^4.3.4",
+    "@angular/core": "^4.3.4",
     "@angular/flex-layout": "^2.0.0-beta.8",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
+    "@angular/forms": "^4.3.4",
+    "@angular/http": "^4.3.4",
     "@angular/material": "^2.0.0-beta.8",
     "@angular/cdk": "^2.0.0-beta.8",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
+    "@angular/platform-browser": "^4.3.4",
+    "@angular/platform-browser-dynamic": "^4.3.4",
+    "@angular/router": "^4.3.4",
     "@nationallibraryofnorway/ngx-mime": "../dist/",
     "classlist.js": "^1.1.20150312",
     "core-js": "^2.4.1",
@@ -39,9 +39,9 @@
     "openseadragon": "^2.2.1"
   },
   "devDependencies": {
-    "@angular/cli": "1.2.7",
-    "@angular/compiler-cli": "^4.0.0",
-    "@angular/language-service": "^4.0.0",
+    "@angular/cli": "1.3.0",
+    "@angular/compiler-cli": "^4.3.4",
+    "@angular/language-service": "^4.3.4",
     "@types/chai": "3.5.2",
     "@types/chai-as-promised": "0.0.30",
     "@types/cucumber": "2.0.1",
