@@ -44,25 +44,25 @@
     "release": "sh scripts/release/release.sh"
   },
   "peerDependencies": {
-    "@angular/core": ">=4.0.0 <5.0.0 || >=4.0.0-beta <5.0.0",
+    "@angular/core": ">=4.3.0 <5.0.0 || >=4.0.0-beta <5.0.0",
     "@angular/material": ">=2.0.0 <3.0.0 || >=2.0.0-beta.8 <3.0.0",
     "@angular/cdk": ">=2.0.0 <3.0.0 || >=2.0.0-beta.8 <3.0.0",
     "@angular/flex-layout": ">=2.0.0 <3.0.0 || >=2.0.0-beta.8 <3.0.0",
     "openseadragon": ">=2.0.0 <3.0.0"
   },
   "devDependencies": {
-    "@angular/common": "~4.2.6",
-    "@angular/compiler": "~4.2.6",
-    "@angular/compiler-cli": "~4.2.6",
-    "@angular/core": "~4.2.6",
-    "@angular/platform-browser": "~4.2.6",
-    "@angular/platform-browser-dynamic": "~4.2.6",
-    "@angular/animations": "~4.2.6",
+    "@angular/common": "4.3.4",
+    "@angular/compiler": "4.3.4",
+    "@angular/compiler-cli": "4.3.4",
+    "@angular/core": "4.3.4",
+    "@angular/platform-browser": "4.3.4",
+    "@angular/platform-browser-dynamic": "4.3.4",
+    "@angular/animations": "4.3.4",
     "@angular/material": "2.0.0-beta.8",
     "@angular/cdk": "2.0.0-beta.8",
-    "@angular/forms": "~4.2.6",
-    "@angular/http": "~4.2.6",
-    "@angular/router": "~4.2.6",
+    "@angular/forms": "4.3.4",
+    "@angular/http": "4.3.4",
+    "@angular/router": "4.3.4",
     "@angular/flex-layout": "^2.0.0-beta.8",
     "hammerjs": "2.0.8",
     "openseadragon": "^2.2.1",
@@ -92,9 +92,10 @@
     "rxjs": "^5.0.1",
     "standard-version": "^4.0.0",
     "systemjs": "^0.19.40",
-    "tslint": "^4.4.2",
+    "tslint": "4.5.1",
     "typescript": "^2.2.1",
     "uglify-es": "^3.0.18",
-    "zone.js": "^0.8.4"
+    "zone.js": "0.8.16",
+    "tslib": "1.7.1"
   }
 }
