@@ -1,8 +1,10 @@
 import { NavbarComponent } from './navbar/navbar.component';
 import { NgModule, Optional, SkipSelf } from '@angular/core';
 import { RouterModule } from '@angular/router';
+import { HTTP_INTERCEPTORS, HttpHandler, HttpClient } from '@angular/common/http';
 
 import { SharedModule } from './../shared/shared.module';
+import { AuthInterceptor } from './auth.interceptor';
 import { SidenavComponent } from './sidenav/sidenav.component';
 import { ManifestService } from './manifest-service/manifest.service';
 
@@ -20,7 +22,8 @@ import { ManifestService } from './manifest-service/manifest.service';
       NavbarComponent
     ],
     providers: [
-        ManifestService
+        ManifestService,
+        { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true },
     ]
 })
 export class CoreModule {
