@@ -1,38 +1,31 @@
-import { CUSTOM_ELEMENTS_SCHEMA, NO_ERRORS_SCHEMA } from '@angular/core';
+import { CUSTOM_ELEMENTS_SCHEMA, DebugElement } from '@angular/core';
 import { async, ComponentFixture, inject, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
-import { DebugElement } from '@angular/core';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
 
 import { ViewerComponent } from './viewer.component';
-import { BaseRequestOptions, ConnectionBackend, Http, HttpModule } from '@angular/http';
-import { MockBackend } from '@angular/http/testing';
 import { IiifService } from '../core/iiif-service/iiif-service';
+import { Manifest } from './../core/models/manifest';
 import { ManifestBuilder } from '../core/builders/manifest.builder';
 import { testManifest } from '../test/testManifest';
 import { Observable } from 'rxjs/Observable';
+
+import 'openseadragon';
 
 describe('ViewerComponent', function () {
   let de: DebugElement;
   let comp: ViewerComponent;
   let fixture: ComponentFixture<ViewerComponent>;
-
+  let spy: any;
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      schemas: [ CUSTOM_ELEMENTS_SCHEMA, NO_ERRORS_SCHEMA ],
+      schemas: [ CUSTOM_ELEMENTS_SCHEMA ],
       imports: [
-        HttpModule
+        HttpClientTestingModule
       ],
       declarations: [ViewerComponent],
       providers: [
-        IiifService,
-        BaseRequestOptions,
-        MockBackend,
-        {
-          provide: Http,
-          useFactory: (backend: ConnectionBackend, defaultOptions: BaseRequestOptions) => {
-            return new Http(backend, defaultOptions);
-          }, deps: [MockBackend, BaseRequestOptions]
-        }
+        IiifService
       ]
     }).compileComponents();
   }));
@@ -48,10 +41,11 @@ describe('ViewerComponent', function () {
   it('should create viewer', inject([IiifService], (iiifService: IiifService) => {
     comp.manifestUri = 'dummyURI';
     const manifest = new ManifestBuilder(testManifest).build();
-    spyOn(iiifService, 'getManifest').and.returnValue(Observable.of(manifest));
+    spy = spyOn(iiifService, 'getManifest').and.returnValue(Observable.of(manifest));
 
     comp.createViewer();
 
     expect(comp.viewer).not.toBeNull();
   }));
+
 });
