@@ -71,19 +71,21 @@ module.exports = function (wallaby) {
         map: {
           '@angular/core/testing': 'npm:@angular/core/bundles/core-testing.umd.js',
           '@angular/common/testing': 'npm:@angular/common/bundles/common-testing.umd.js',
+          '@angular/common/http/testing': 'npm:@angular/common/bundles/common-http-testing.umd.js',
           '@angular/compiler/testing': 'npm:@angular/compiler/bundles/compiler-testing.umd.js',
           '@angular/platform-browser/testing': 'npm:@angular/platform-browser/bundles/platform-browser-testing.umd.js',
           '@angular/platform-browser-dynamic/testing': 'npm:@angular/platform-browser-dynamic/bundles/platform-browser-dynamic-testing.umd.js',
           '@angular/http/testing': 'npm:@angular/http/bundles/http-testing.umd.js',
           '@angular/router/testing': 'npm:@angular/router/bundles/router-testing.umd.js',
           '@angular/forms/testing': 'npm:@angular/forms/bundles/forms-testing.umd.js',
+          'openseadragon': 'npm:openseadragon/build/openseadragon/openseadragon.min.js',
         },
       });
 
       var promises = [
         Promise.all([
           System.import('@angular/core/testing'),
-          System.import('@angular/platform-browser-dynamic/testing')
+          System.import('@angular/platform-browser-dynamic/testing'),
         ])
 
           .then(function (providers) {
